requirements.txt
pika
python-dotenv
mysql-connector-python